Move Attribute class into its own sources
