+++ src/creature/attributes.h
@@ -24,6 +24,9 @@
 #ifndef LCS_CREATURE_ATTRIBUTES_H_H
 #define LCS_CREATURE_ATTRIBUTES_H_H
 
+#include <string>
+
+
 enum CreatureAttribute
 {
    ATTRIBUTE_STRENGTH,
@@ -32,8 +35,36 @@ enum CreatureAttribute
    ATTRIBUTE_AGILITY,
    ATTRIBUTE_HEALTH,
    ATTRIBUTE_CHARISMA,
-   ATTRIBUTE_HEART,
-   ATTNUM
+   ATTRIBUTE_HEART
+};
+
+constexpr std::size_t ATTNUM = 7;
+constexpr int         MIN_ATTR_LEVEL = 0;
+constexpr int         MAX_ATTR_LEVEL = 99;
+
+
+class Attribute
+{
+public:
+   Attribute()
+   { }
+
+   Attribute(std::string const& inputXml);
+
+   std::string
+   showXml() const;
+
+   void
+   set_type(int attribute_type)
+   { attribute = attribute_type; }
+
+   static std::string get_name(int attribute_type);
+
+public:
+   int value;
+
+private:
+   int attribute;
 };
 
 #endif /* LCS_CREATURE_ATTRIBUTES_H_H */
